+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Dev
+## 8.2.1
 * adopt to Playwright 1.51, improve stack trace parsing.
 * hide "_" prefixed attachments.